// OmkarFCUBEAPP/src/app/models/drivermodel.ts
export class Drivermodel {
    driverMasterID: string = "";
    driverName: string = "";
    fatherName: string = "";
    dateOfBirth: string = "";
    age: string = "";
    introBy: string = "";
    introByMobileNo: string = "";
    dateOfAppoint: string = "";
    licenseNo: string = "";
    licValidUpto: string = "";
    licenseIssuAuth: string = "";
    isHazardousLicYN: string = "";
    hazLicenseIssuAuth: string = "";
    hazardousLicNo: string = "";
    hazLicValidUpto: string = "";
    bloodGroup: string = "";
    typeOfLicence: string = "";
    driverMobile1: string = "";
    driverMobile2: string = "";
    temporaryAddr: string = "";
    temporaryAddrCity: string = "";
    temporaryAddrPin: string = "";
    tempAddPhone: string = "";
    permanentAddr: string = "";
    permanentAddrCity: string = "";
    permanentAddrPin: string = "";
    permAddPhone: string = "";
    driverAadharNo: string = "";
    previousExpDetails: string = "";
    previousExpYears: string = "";
    isActive: string = "";
    inActiveDate: string = "";
    removedYN: string = "";
    removedDate: string = "";
    remarks: string = "";
    groupName: string = "";
    driverAcct: string = "";
    drPhoto: string = "";
    attachDrLic: string = "";
    attachDrHazLic: string = "";
    attachDrAadhar: string = "";
    attachDrTempAddProof: string = "";
    attachDrPermAddProof: string = "";
    attachDrBankPassBook: string = "";
    bankName: string = "";
    drBankAccountName: string = "";
    bankAcNo: string = "";
    bankBranch: string = "";
    bankIfsCode: string = "";
    bankAccountStatus: string = "";
    createdBy: string = "";
    modifiedBy: string = "";
    loggedInUser: string = "";
    deleteFlag: string = "";
}